% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bats.R
\name{bats}
\alias{bats}
\title{Get BATS (Bermuda Atlantic Time-series Study) data}
\usage{
bats(dataset = "production")
}
\arguments{
\item{dataset}{(character) One of production, zooplankton, or flux}
}
\description{
Get BATS (Bermuda Atlantic Time-series Study) data
}
\details{
Only these three data sets for now. Others may follow.
}
\examples{
\dontrun{
# Production
(res <- bats("production"))
res$meta
res$vars

# Zooplankton
(res <- bats("zooplankton"))
res$meta
res$vars

# Flux
(res <- bats("flux"))
res$meta
res$vars
}
}
\references{
\url{http://bats.bios.edu/}
}
