% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bats_production.R
\name{bats_production}
\alias{bats_production}
\title{Get production data from BATS.}
\usage{
bats_production()
}
\description{
Get production data from BATS.
}
\examples{
\donttest{
(res <- bats_production())
res$meta
res$vars
}
}

