% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcofi.R
\name{calcofi}
\alias{calcofi}
\title{Get California Cooperative Oceanic Fisheries Investigations (CALCOFI) data.}
\usage{
calcofi(dataset = "hydro_cast", path = "~/.ots/calcofi", overwrite = TRUE)
}
\arguments{
\item{dataset}{A dataset code name, one of hydro, macrozoo, or X}

\item{path}{A path to store the files, Default: \code{~/.ots/kelp}}

\item{overwrite}{(logical) To overwrite the path to store files in or not, Default: TRUE.}
}
\description{
FIXME: need to support other data sets in CALCOFI
}
\examples{
\dontrun{
# hydro cast data
(res <- calcofi('hydro_cast'))

# hydro bottle data
(res <- calcofi('hydro_bottle'))

library('dplyr')
res$data \%>\%
 tbl_df \%>\%
 select(-cruz_sta) \%>\%
 group_by(year) \%>\%
 summarise(avg_wind_spd = mean(wind_spd, na.rm = TRUE)) \%>\%
 arrange(desc(avg_wind_spd)) \%>\%
 head
}
}

