% Generated by roxygen2 (4.0.2): do not edit by hand
\name{kelp}
\alias{kelp}
\alias{kelp_datasets}
\title{Get Kelp forest community data.}
\usage{
kelp(which = "benthic_cover", path = "~/.ots/kelp", overwrite = TRUE)

kelp_datasets()
}
\arguments{
\item{which}{A dataset code name, see \code{\link{kelp_datasets}}}

\item{path}{A path to store the files, Default: \code{~/.ots/kelp}}

\item{overwrite}{(logical) To overwrite the path to store files in or not, Default: TRUE.}
}
\description{
Get Kelp forest community data.
}
\details{
After one download o the dataset, you won't have to download the data again.
}
\examples{
\donttest{
# list of datasets
kelp_datasets()

# get data
(res <- kelp("benthic_cover"))
head(res$headers)
head(res$vars)
res$citation

(res <- kelp("benthic_density"))
(res <- kelp("fish_density"))
(res <- kelp("fish_size"))
(res <- kelp("invert_size"))
head(res$headers)
(res <- kelp("subtidal"))
(res <- kelp("rdfc"))
(res <- kelp("kelp_size"))
(res <- kelp("kelp_supp_dens"))
(res <- kelp("art_recruit"))
}
}

