% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ots-package.R
\docType{package}
\name{ots-package}
\alias{ots}
\alias{ots-package}
\title{ots, an R client for various ocean time series datasets.}
\description{
ots, an R client for various ocean time series datasets.
}

