% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uopg.R
\name{uopg}
\alias{uopg}
\title{Get data from the Upper Ocean Proccesses Group (UOPG).}
\usage{
uopg(
  dataset = "smile",
  type = "meteorology",
  path = "~/.ots/uopg",
  overwrite = TRUE
)
}
\arguments{
\item{dataset}{A dataset code name, one of arabian_sea, asrex_91, asrex_93, biowatt, cmo,
coare, coop, fasinex, lotus, mlml89, mlml91, sesmoor, smile, or subduction. Or their
unique abbreviations.}

\item{type}{A data type, one of meteorology, water_velocity, temperature, or salinity. Or their
unique abbreviations.}

\item{path}{A path to store the files, Default: \code{~/.ots/uopg}}

\item{overwrite}{(logical) To overwrite the path to store files in or not, Default: TRUE.}
}
\description{
Get data from the Upper Ocean Proccesses Group (UOPG).
}
\details{
We download NetCDF files from UOPG, using \pkg{ncdf} to parse the data
to a data.frame.
}
\examples{
\dontrun{
# Smile dataaets
(met <- uopg(dataset = 'smile', type = "meteorology"))
(sal <- uopg(dataset = 'smile', type = "salinity"))
(water <- uopg(dataset = 'smile', type = "water"))
(temp <- uopg(dataset = 'smile', type = "temp"))

# biowatt dataaets
(biowatt_met <- uopg(dataset = 'biowatt', type = "meteorology"))
biowatt_met$data$`1`

# lotus datasets
(lotus_met <- uopg(dataset = 'lotus', type = "meteorology"))
lotus_met$data$lotus
lotus_met$metadata

# coare datasets
(coare_sal <- uopg(dataset = 'coare', type = "salinity"))
}
}
\references{
\url{http://uop.whoi.edu/archives/dataarchives.html}
}
