% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bats_zooplankton.R
\name{bats_zooplankton}
\alias{bats_zooplankton}
\title{Get zooplankton data from BATS.}
\usage{
bats_zooplankton()
}
\description{
Get zooplankton data from BATS.
}
\examples{
\donttest{
(res <- bats_zooplankton())
res$meta
res$vars
}
}

