% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bats_zooplankton}
\alias{bats_zooplankton}
\alias{print.zooplankton}
\title{Get zooplankton data from BATS.}
\usage{
bats_zooplankton()

\method{print}{zooplankton}(x, ..., n = 10)
}
\description{
Get zooplankton data from BATS.
}
\examples{
\donttest{
res <- bats_zooplankton()
}
}

